app/requirements.txt
alembic @ file:///home/conda/feedstock_root/build_artifacts/alembic_1763216776648/work
black==25.11.0
blinker @ file:///home/conda/feedstock_root/build_artifacts/blinker_1731096409132/work
certifi==2025.11.12
charset-normalizer==3.4.4
click @ file:///home/conda/feedstock_root/build_artifacts/click_1763248546991/work
defusedxml==0.7.1
Flask @ file:///home/conda/feedstock_root/build_artifacts/flask_1755674618058/work
flask-cors==6.0.1
Flask-Migrate @ file:///home/conda/feedstock_root/build_artifacts/flask-migrate_1736546704161/work
Flask-SQLAlchemy @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_flask-sqlalchemy_1733868589/work
fonttools==4.61.0
fpdf2==2.8.5
greenlet @ file:///home/conda/feedstock_root/build_artifacts/greenlet_1756752057234/work
gunicorn @ file:///home/conda/feedstock_root/build_artifacts/gunicorn_1762506418880/work
idna==3.11
importlib_metadata @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_importlib-metadata_1747934053/work
iniconfig==2.3.0
itsdangerous @ file:///home/conda/feedstock_root/build_artifacts/itsdangerous_1733308265247/work
Jinja2 @ file:///home/conda/feedstock_root/build_artifacts/jinja2_1741263328855/work
Mako @ file:///home/conda/feedstock_root/build_artifacts/mako_1744317760971/work
MarkupSafe @ file:///home/conda/feedstock_root/build_artifacts/markupsafe_1759055168201/work
mypy_extensions==1.1.0
numpy @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_numpy_1763351006/work/dist/numpy-2.3.5-cp311-cp311-linux_x86_64.whl#sha256=321966b39af6fd2451eac0fb457d64b1c9253896a6d48f621f8b24194d97e91a
packaging @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_packaging_1745345660/work
pandas @ file:///home/conda/feedstock_root/build_artifacts/pandas_1759265546765/work
pathspec==0.12.1
pillow==12.0.0
platformdirs==4.5.0
pluggy==1.6.0
psycopg2 @ file:///home/conda/feedstock_root/build_artifacts/psycopg2-split_1763595848516/work
psycopg2-binary @ file:///home/conda/feedstock_root/build_artifacts/psycopg2-split_1750255432256/work/psycopg2-binary
Pygments==2.19.2
pytest==9.0.1
python-dateutil @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_python-dateutil_1751104122/work
python-dotenv @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_python-dotenv_1761503229/work
pytokens==0.3.0
pytz @ file:///home/conda/feedstock_root/build_artifacts/pytz_1742920838005/work
PyYAML @ file:///home/conda/feedstock_root/build_artifacts/pyyaml_1758891829620/work
requests==2.32.5
six @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_six_1753199211/work
SQLAlchemy @ file:///home/conda/feedstock_root/build_artifacts/sqlalchemy_1760114468284/work
tomli @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_tomli_1760014679/work
typing_extensions @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_typing_extensions_1756220668/work
tzdata @ file:///home/conda/feedstock_root/build_artifacts/python-tzdata_1742745135198/work
urllib3==2.5.0
watchdog @ file:///home/conda/feedstock_root/build_artifacts/watchdog_1763021587200/work
Werkzeug @ file:///home/conda/feedstock_root/build_artifacts/werkzeug_1733160440960/work
zipp @ file:///home/conda/feedstock_root/build_artifacts/zipp_1749421620841/work
